requirements.txt.txt
fastapi
uvicorn
httpx
pillow
python-multipart
